no se si lo que he petado o que pero hay cosas que ya no van

+++ P2/primera_parte/includes/redNeuronal.h
@@ -73,12 +73,12 @@ int actualizaSalida(redNeuronal* red, double (*fActualizacion)(neurona*), double
 ***	Tener ojo con usar una actulizacion bipolar
 ***
 **/
-int actualizaPesosPerceptron(redNeuronal* red, int* t);
+double actualizaPesosPerceptron(redNeuronal* red, int* t);
 
-int actualizaPesosAdaline(redNeuronal* red, int* t);
+double actualizaPesosAdaline(redNeuronal* red, int* t);
 
 
-int actualizaPesosRetropropagacion(redNeuronal* red, int* t);
+double actualizaPesosRetropropagacion(redNeuronal* red, int* t);
 
 /**
 ***
@@ -106,7 +106,7 @@ redNeuronal* redTrain(int tentrada,datos* data,
 					redNeuronal* (*fini)(int, int, int, double),
 					int (*fsalida) (redNeuronal*, double (*fActualizacion)(neurona*), double*),
 					int (*fParada) (redNeuronal*),
-					int (*fPesos) (redNeuronal*, int*),
+					double (*fPesos) (redNeuronal*, int*),
 					double (*fActualizacion)(neurona*),					
 					int nentreada, int nsalida, int noculta, double tasa);
 /**1
